Move get-job-list endpoint to job.py
